requirements.txt
# =============================================================================
# Multi-Model AI Platform — Python Dependencies
# =============================================================================

# --- API Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# --- Image Generation (ComfyUI client) ---
aiohttp==3.11.11
websockets==14.1
Pillow==11.1.0

# --- Video Generation (Wan 2.1) ---
diffusers==0.33.0
transformers==4.48.1
accelerate==1.3.0
safetensors==0.5.2
imageio==2.37.0
imageio-ffmpeg==0.5.1

# --- Text Generation (Mixtral GGUF) ---
llama-cpp-python==0.3.8

# --- Audio Generation (XTTS v2) ---
TTS==0.22.0

# --- Core ML ---
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1

# --- Utilities ---
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
numpy==1.26.4
scipy==1.15.1
